+++ TongQuocKhanh_22677951_JPA/src/main/java/iuh/fit/tongquockhanh_22677951_jpa/repository/DepartmentRepository.java
@@ -3,4 +3,4 @@ package iuh.fit.tongquockhanh_22677951_jpa.repository;
 import iuh.fit.tongquockhanh_22677951_jpa.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepository extends JpaRepository<Department, Long> { }
+public interface DepartmentRepository extends JpaRepository<Department, Integer> { }
